add heat to systems
