+++ modules/core/src/configurations/dragonfly-sf-22.ts
@@ -37,6 +37,8 @@ export const dragonflyReactor = {
     maxAfterBurnerFuel: 5000,
     afterBurnerCharge: 20,
     afterBurnerEnergyCost: 0.07,
+    energyHeatEPMThreshold: 5,
+    energyHeat: 0.1,
     damage50: 20,
 };
 export const dragonflyProperties = {
