+++ modules/core/src/ship/docking.ts
@@ -2,6 +2,7 @@ import { DesignState, defectible } from './system';
 import { Schema, type } from '@colyseus/schema';
 
 import { EPSILON } from '../logic';
+import { MAX_SYSTEM_HEAT } from './heat-manager';
 import { number2Digits } from '../number-field';
 import { range } from '../range';
 import { tweakable } from '../tweakable';
@@ -40,6 +41,11 @@ export class Docking extends Schema {
 
     @number2Digits
     public energyPerMinute = 0;
+
+    @range([0, MAX_SYSTEM_HEAT])
+    @tweakable('number')
+    @number2Digits
+    public heat = 0;
 
     @type(DockingDesignState)
     design = new DockingDesignState();
