+++ modules/core/src/ship/magazine.ts
@@ -1,7 +1,7 @@
+import { DesignState, defectible } from './system';
 import { Schema, type } from '@colyseus/schema';
 
-import { DesignState } from './system';
-import { defectible } from './system';
+import { MAX_SYSTEM_HEAT } from './heat-manager';
 import { number2Digits } from '../number-field';
 import { range } from '../range';
 import { tweakable } from '../tweakable';
@@ -36,6 +36,11 @@ export class Magazine extends Schema {
 
     @number2Digits
     public energyPerMinute = 0;
+
+    @range([0, MAX_SYSTEM_HEAT])
+    @tweakable('number')
+    @number2Digits
+    public heat = 0;
 
     @type(MagazineDesignState)
     design = new MagazineDesignState();
